Refs #10409 - Rails 4 compatible changes with Rails 3
This includes some changes that can be applied to Rails 3 without
any issue and don't follow any particular pattern/deprecation, In
general these are minor things we have to change. Other PRs with
retrocompatible changes that aim to fix particular deprecations
have been submitted separately.

Sorry for the not-very-descriptive commit!