Fixes #11986 - Specify PUT method on form_for tags
Rails 4 will default to PATCH which we have not implemented in our
routes. This will cause at least one failure on taxonomies integration
tests. It's the only one our test suite catches, however I'd expect
other forms to fail on this when tested manually. We can keep this
issue number as a reference and add commits with integration tests
& fixes where needed.